Dialogs and Toasts

+++ app/favorites/favorites.component.ts
@@ -5,6 +5,8 @@ import { ListViewEventData, RadListView } from 'nativescript-ui-listview';
 import { RadListViewComponent } from 'nativescript-ui-listview/angular';
 import { ObservableArray } from 'tns-core-modules/data/observable-array';
 import { View } from 'tns-core-modules/ui/core/view';
+import { confirm } from "ui/dialogs";
+import { Toasty } from 'nativescript-toasty';
 
 @Component({
     selector: 'app-favorites',
@@ -22,20 +24,48 @@ export class FavoritesComponent implements OnInit {
     constructor(
         private favoriteservice: FavoriteService,
         @Inject('baseURL') private baseURL
-        ) {
+    ) {
     }
 
     ngOnInit() {
         this.favoriteservice.getFavorites()
-            .subscribe(favorites => this.favorites = new ObservableArray(favorites),
+            .subscribe(favorites => {
+                this.favorites = new ObservableArray(favorites);
+                console.log("Total Favorite => " + this.favorites.length.toString());
+            },
                 errmess => this.errMess = errmess);
     }
 
     deleteFavorite(id: string) {
-        this.favoriteservice.deleteFavorite(id)
-            .subscribe(favorites => this.favorites = new ObservableArray(favorites),
-                errmess => this.errMess = errmess);
-    }
+        console.log('delete', id);
+    
+        let options = {
+            title: "Confirm Delete",
+            message: 'Do you want to delete Dish '+ id,
+            okButtonText: "Yes",
+            cancelButtonText: "No",
+            neutralButtonText: "Cancel"
+        };
+    
+        confirm(options).then((result: boolean) => {
+            if(result) {
+    
+              this.favorites = null;
+    
+              this.favoriteservice.deleteFavorite(id)
+                  .subscribe(favorites => { 
+                    const toast = new Toasty("Deleted Dish "+ id, "short", "bottom");
+                    toast.show();
+                    this.favorites = new ObservableArray(favorites);
+                  },
+                  errmess => this.errMess = errmess);
+            }
+            else {
+              console.log('Delete cancelled');
+            }
+        });
+    
+      }
 
     public onCellSwiping(args: ListViewEventData) {
         var swipeLimits = args.data.swipeLimits;
